mm/gup: Add an option for obtaining an exclusive pin
Introduce the ability to obtain an exclusive long-term pin on a
page. This exclusive pin can only be held if there are no other
pins on the page, regular, or exclusive. Moreover, once this pin
is held, no other pins can be grabbed until the exclusive pin is
released.

This pin is grabbed using the (new) FOLL_EXCLUSIVE flag, and is
gated by the EXCLUSIVE_PIN configuration option.

Similar to how the normal GUP pin is obtain, the exclusive PIN
overloads the _refcount field for normal pages, or the _pincount
field for large pages. It appropriates bit 30 of these two
fields, which still allows the detection of overflows into bit
31. It does however, half the number of potential normals pins
for a page.

In order to avoid the possibility of COWing such a page, once an
exclusive pin has been obtained, it's marked as AnonExclusive.

Co-Developed-by: Elliot Berman <quic_eberman@quicinc.com>
Signed-off-by: Elliot Berman <quic_eberman@quicinc.com>
Signed-off-by: Fuad Tabba <tabba@google.com>